% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle-data.R
\name{convert_wide_mat_to_long_df}
\alias{convert_wide_mat_to_long_df}
\title{Converts longitudinal data from "wide" to "long" format
(assumes that the incoming dataset is stored as a matrix object)}
\usage{
convert_wide_mat_to_long_df(wide_mat, var_name)
}
\arguments{
\item{wide_mat}{Matrix of dimension N_individuals x K_timepoints}

\item{var_name}{Variable name for the data values}
}
\value{
Data frame containing longitudinal data in long format
}
\description{
Converts longitudinal data from "wide" to "long" format
(assumes that the incoming dataset is stored as a matrix object)
}
