% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longsim-package.R
\docType{package}
\name{longsim-package}
\alias{longsim}
\alias{longsim-package}
\title{longsim: Longitudinal Data Simulations for Power Calculations and Conceptual Visualization}
\description{
Longitudinal data simulations for power calculations and conceptual visualization.
}
\author{
\strong{Maintainer}: Kenny Westerman \email{ken.e.westerman@gmail.com} (\href{https://orcid.org/0000-0001-7619-1868}{ORCID})

}
\keyword{internal}
