% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle-data.R
\name{merge_variables}
\alias{merge_variables}
\title{Converts a list of matrices into a single dataset in long format}
\usage{
merge_variables(mat_list)
}
\arguments{
\item{mat_list}{List of matrices, all of dimension N_individuals x K_timepoints}
}
\value{
Data frame in long format containing a column for each variable in
the input matrix list
}
\description{
Converts a list of matrices into a single dataset in long format
}
