% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_covariate}
\alias{simulate_covariate}
\title{Simulates a matrix of time-varying covariates}
\usage{
simulate_covariate(N, K, icc_C, var_e_C, frac_missing = 0)
}
\arguments{
\item{N}{Number of individuals}

\item{K}{Number of measurement timepoints (assumes equal number per individual for now)}

\item{icc_C}{Intraclass correlation coefficient for the covariate (clustering within an individual across time)}

\item{var_e_C}{Error variance for the covariate}

\item{frac_missing}{Fraction of all values set to missing (MCAR)}
}
\value{
Covariate matrix of dimension N_individuals x K timepoints
}
\description{
Simulates a matrix of time-varying covariates
}
