% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_genotype}
\alias{simulate_genotype}
\title{Simulates a matrix of genotypes for a single hypothetical variant}
\usage{
simulate_genotype(N, K, maf = 0.25)
}
\arguments{
\item{N}{Number of individuals}

\item{K}{Number of measurement timepoints (assumes equal number per individual for now)}

\item{maf}{Genotype minor allele frequency}
}
\value{
Genotype matrix of dimension N_individuals x K timepoints
}
\description{
Simulates a matrix of genotypes for a single hypothetical variant
}
