% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_long_data}
\alias{simulate_long_data}
\title{Simulate a longitudinal dataset}
\usage{
simulate_long_data(
  N = 100,
  K = 3,
  maf = 0.25,
  icc_C = 0.5,
  var_e_C = 1,
  beta_CE = 0,
  icc_E = 0.5,
  var_e_E = 1,
  beta_tY = -1,
  beta_GY = 0.5,
  beta_CY = 0,
  beta_EY = 1,
  icc_Y = 0.8,
  var_e_Y = 1,
  frac_missing = 0
)
}
\arguments{
\item{N}{Number of individuals}

\item{K}{Number of measurement timepoints (assumes equal number per individual for now)}

\item{maf}{Genotype minor allele frequency}

\item{icc_C}{Intraclass correlation coefficient for the covariate (clustering within an individual across time)}

\item{var_e_C}{Error variance for the covariate}

\item{beta_CE}{Confounder effect size on E}

\item{icc_E}{Intraclass correlation coefficient for the exposure (clustering within an individual across time)}

\item{var_e_E}{Error variance for the exposure}

\item{beta_tY}{Time effect on Y}

\item{beta_GY}{Genotype effect on Y}

\item{beta_CY}{Covariate effect on Y}

\item{beta_EY}{Exposure effect on Y}

\item{icc_Y}{Intraclass correlation coefficient for Y (clustering within an individual across time)}

\item{var_e_Y}{Error variance for Y}

\item{frac_missing}{Fraction of all values set to missing (MCAR)}
}
\value{
Data frame in long format containing fields for ID (id), timepoint label (timept),
outcome (Y), measurement time (t), genotype (G),exposure (E), and covariate (C)
}
\description{
Simulate a longitudinal dataset
}
