% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_outcome}
\alias{simulate_outcome}
\title{Simulates a longitudinal outcome variable}
\usage{
simulate_outcome(matrix_list, beta_list, icc_Y, var_e_Y, frac_missing = 0)
}
\arguments{
\item{matrix_list}{List of matrices affecting the mean of Y}

\item{beta_list}{List of effect sizes associated with the matrices in matrix_list}

\item{icc_Y}{Intraclass correlation coefficient for Y (clustering within an individual across time)}

\item{var_e_Y}{Error variance for Y}

\item{frac_missing}{Fraction of all values set to missing (MCAR)}
}
\value{
Outcome matrix of dimension N_individuals x K timepoints
}
\description{
Simulates a longitudinal outcome variable
}
