% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_outcome}
\alias{simulate_outcome}
\title{Simulates a longitudinal outcome variable}
\usage{
simulate_outcome(
  t_mat,
  G_mat,
  E_mat,
  beta_tY,
  beta_GY,
  beta_EY,
  icc_Y,
  var_e_Y
)
}
\arguments{
\item{t_mat}{Matrix of times for each measurement}

\item{G_mat}{Matrix of genotypes (N x K)}

\item{E_mat}{Matrix of exposures (N x K)}

\item{beta_tY}{Linear slope of time effect on Y}

\item{beta_GY}{Genotype effect size on Y}

\item{beta_EY}{Exposure effect size on Y}

\item{icc_Y}{Intraclass correlation coefficient for Y (clustering within an individual across time)}

\item{var_e_Y}{Error variance for Y}
}
\value{
Outcome matrix of dimension N_individuals x K timepoints
}
\description{
Simulates a longitudinal outcome variable
}
